Refactored player listener and renamed main class
